requirements.txt
pillow
tk
pyscreenshot
pyautogui
pytesseract
pyperclip